+++ lib/influxdb2/client/default_api.rb
@@ -83,6 +83,8 @@ module InfluxDB2
         else
           raise InfluxError.from_response(response)
         end
+      rescue *InfluxError::HTTP_ERRORS => error
+        raise InfluxError.from_error(error)
       ensure
         http.finish if http.started?
       end