+++ test/influxdb/write_api_batching_test.rb
@@ -187,6 +187,65 @@ class WriteApiBatchingTest < MiniTest::Test
                      'h2o_feet,location=coyote_creek level\\ water_level=3.0 3')
   end
 
+  def test_jitter_interval
+    @client.close!
+
+    @client = InfluxDB2::Client.new('http://localhost:9999',
+                                    'my-token',
+                                    bucket: 'my-bucket',
+                                    org: 'my-org',
+                                    precision: InfluxDB2::WritePrecision::NANOSECOND,
+                                    use_ssl: false)
+
+    @write_options = InfluxDB2::WriteOptions.new(write_type: InfluxDB2::WriteType::BATCHING,
+                                                 batch_size: 2, flush_interval: 5_000, jitter_interval: 2_000)
+    @write_client = @client.create_write_api(write_options: @write_options)
+
+    stub_request(:post, 'http://localhost:9999/api/v2/write?bucket=my-bucket&org=my-org&precision=ns')
+      .to_return(status: 204)
+
+    request = "h2o_feet,location=coyote_creek water_level=1.0 1\n" \
+               'h2o_feet,location=coyote_creek water_level=2.0 2'
+
+    @write_client.write(data: ['h2o_feet,location=coyote_creek water_level=1.0 1',
+                               'h2o_feet,location=coyote_creek water_level=2.0 2'])
+
+    sleep(0.05)
+
+    assert_requested(:post, 'http://localhost:9999/api/v2/write?bucket=my-bucket&org=my-org&precision=ns',
+                     times: 0, body: request)
+
+    sleep(2)
+
+    assert_requested(:post, 'http://localhost:9999/api/v2/write?bucket=my-bucket&org=my-org&precision=ns',
+                     times: 1, body: request)
+  end
+end
+
+class WriteApiRetryStrategyTest < MiniTest::Test
+  def setup
+    WebMock.disable_net_connect!
+
+    @write_options = InfluxDB2::WriteOptions.new(write_type: InfluxDB2::WriteType::BATCHING,
+                                                 batch_size: 2, flush_interval: 5_000, retry_interval: 2_000)
+    @client = InfluxDB2::Client.new('http://localhost:9999',
+                                    'my-token',
+                                    bucket: 'my-bucket',
+                                    org: 'my-org',
+                                    precision: InfluxDB2::WritePrecision::NANOSECOND,
+                                    use_ssl: false)
+
+    @write_client = @client.create_write_api(write_options: @write_options)
+  end
+
+  def teardown
+    @client.close!
+
+    assert_equal true, @write_client.closed
+
+    WebMock.reset!
+  end
+
   def test_retry_interval_by_config
     error_body = '{"code":"temporarily unavailable","message":"Token is temporarily over quota. '\
                  'The Retry-After header describes when to try the write again."}'
@@ -211,7 +270,7 @@ class WriteApiBatchingTest < MiniTest::Test
     assert_requested(:post, 'http://localhost:9999/api/v2/write?bucket=my-bucket&org=my-org&precision=ns',
                      times: 1, body: request)
 
-    sleep(1)
+    sleep(5)
 
     assert_requested(:post, 'http://localhost:9999/api/v2/write?bucket=my-bucket&org=my-org&precision=ns',
                      times: 2, body: request)
@@ -256,37 +315,234 @@ class WriteApiBatchingTest < MiniTest::Test
                      times: 2, body: request)
   end
 
-  def test_jitter_interval
-    @client.close!
+  def test_max_retries
+    error_body = '{"code":"temporarily unavailable","message":"Server is temporarily unavailable to accept writes. '\
+                 'The Retry-After header describes when to try the write again."}'
 
-    @client = InfluxDB2::Client.new('http://localhost:9999',
-                                    'my-token',
-                                    bucket: 'my-bucket',
-                                    org: 'my-org',
-                                    precision: InfluxDB2::WritePrecision::NANOSECOND,
-                                    use_ssl: false)
-
-    @write_options = InfluxDB2::WriteOptions.new(write_type: InfluxDB2::WriteType::BATCHING,
-                                                 batch_size: 2, flush_interval: 5_000, jitter_interval: 2_000)
-    @write_client = @client.create_write_api(write_options: @write_options)
+    headers = { 'X-Platform-Error-Code' => 'temporarily unavailable' }
 
     stub_request(:post, 'http://localhost:9999/api/v2/write?bucket=my-bucket&org=my-org&precision=ns')
-      .to_return(status: 204)
+      .to_return(status: 429, headers: headers, body: error_body).then # retry
+      .to_return(status: 429, headers: headers, body: error_body).then # retry
+      .to_return(status: 429, headers: headers, body: error_body).then # retry
+      .to_return(status: 429, headers: headers, body: error_body).then # retry
+      .to_return(status: 429, headers: headers, body: error_body) # not called
 
-    request = "h2o_feet,location=coyote_creek water_level=1.0 1\n" \
-               'h2o_feet,location=coyote_creek water_level=2.0 2'
+    point = InfluxDB2::Point.new(name: 'h2o')
+                            .add_tag('location', 'europe')
+                            .add_field('level', 2.0)
 
-    @write_client.write(data: ['h2o_feet,location=coyote_creek water_level=1.0 1',
-                               'h2o_feet,location=coyote_creek water_level=2.0 2'])
+    request = 'h2o,location=europe level=2.0'
 
-    sleep(0.05)
+    write_options = InfluxDB2::WriteOptions.new(write_type: InfluxDB2::WriteType::BATCHING,
+                                                batch_size: 1, retry_interval: 2_000, max_retries: 3,
+                                                max_retry_delay: 5_000, exponential_base: 2)
+
+    error = assert_raises InfluxDB2::InfluxError do
+      @client.create_write_api(write_options: write_options).write(data: point)
+
+      sleep(0.5)
+
+      assert_requested(:post, 'http://localhost:9999/api/v2/write?bucket=my-bucket&org=my-org&precision=ns',
+                       times: 1, body: request)
+
+      sleep(2)
+
+      assert_requested(:post, 'http://localhost:9999/api/v2/write?bucket=my-bucket&org=my-org&precision=ns',
+                       times: 2, body: request)
+
+      sleep(4)
+
+      assert_requested(:post, 'http://localhost:9999/api/v2/write?bucket=my-bucket&org=my-org&precision=ns',
+                       times: 3, body: request)
+
+      sleep(5)
+
+      assert_requested(:post, 'http://localhost:9999/api/v2/write?bucket=my-bucket&org=my-org&precision=ns',
+                       times: 4, body: request)
+
+      sleep(5)
+
+      assert_requested(:post, 'http://localhost:9999/api/v2/write?bucket=my-bucket&org=my-org&precision=ns',
+                       times: 4, body: request)
+
+      sleep(5)
+    end
+
+    assert_equal('429', error.code)
 
     assert_requested(:post, 'http://localhost:9999/api/v2/write?bucket=my-bucket&org=my-org&precision=ns',
-                     times: 0, body: request)
+                     times: 4, body: request)
+  end
+
+  def test_influx_exception
+    error_body = '{"code":"invalid","message":"unable to parse '\
+                 '\'h2o_feet, location=coyote_creek water_level=1.0 1\': missing tag key"}'
+
+    stub_request(:any, 'http://localhost:9999/api/v2/write?bucket=my-bucket&org=my-org&precision=ns')
+      .to_return(status: 400, headers: { 'X-Platform-Error-Code' => 'invalid' }, body: error_body)
+
+    write_options = InfluxDB2::WriteOptions.new(write_type: InfluxDB2::WriteType::BATCHING,
+                                                batch_size: 1, retry_interval: 2_000, max_retries: 3,
+                                                max_retry_delay: 5_000, exponential_base: 2)
+
+    error = assert_raises InfluxDB2::InfluxError do
+      @client.create_write_api(write_options: write_options).write(data: 'h2o,location=west value=33i 15')
+
+      sleep(1)
+    end
+
+    assert_equal '400', error.code
+    assert_equal 'invalid', error.reference
+    assert_equal "unable to parse 'h2o_feet, location=coyote_creek water_level=1.0 1': missing tag key", error.message
+  end
+
+  def test_max_retries_by_header
+    error_body = '{"code":"temporarily unavailable","message":"Server is temporarily unavailable to accept writes. '\
+                 'The Retry-After header describes when to try the write again."}'
+
+    headers = { 'X-Platform-Error-Code' => 'temporarily unavailable', 'Retry-After' => '3' }
+
+    stub_request(:post, 'http://localhost:9999/api/v2/write?bucket=my-bucket&org=my-org&precision=ns')
+      .to_return(status: 429, headers: headers, body: error_body).then # retry
+      .to_return(status: 429, headers: headers, body: error_body).then # retry
+      .to_return(status: 429, headers: headers, body: error_body).then # retry
+      .to_return(status: 429, headers: headers, body: error_body).then # retry
+      .to_return(status: 429, headers: headers, body: error_body) # not called
+
+    point = InfluxDB2::Point.new(name: 'h2o')
+                            .add_tag('location', 'europe')
+                            .add_field('level', 2.0)
+
+    request = 'h2o,location=europe level=2.0'
+
+    write_options = InfluxDB2::WriteOptions.new(write_type: InfluxDB2::WriteType::BATCHING,
+                                                batch_size: 1, retry_interval: 2_000, max_retries: 3,
+                                                max_retry_delay: 5_000, exponential_base: 2)
+
+    error = assert_raises InfluxDB2::InfluxError do
+      @client.create_write_api(write_options: write_options).write(data: point)
+
+      sleep(0.5)
+
+      assert_requested(:post, 'http://localhost:9999/api/v2/write?bucket=my-bucket&org=my-org&precision=ns',
+                       times: 1, body: request)
+
+      sleep(3)
+
+      assert_requested(:post, 'http://localhost:9999/api/v2/write?bucket=my-bucket&org=my-org&precision=ns',
+                       times: 2, body: request)
+
+      sleep(3)
+
+      assert_requested(:post, 'http://localhost:9999/api/v2/write?bucket=my-bucket&org=my-org&precision=ns',
+                       times: 3, body: request)
+
+      sleep(3)
+
+      assert_requested(:post, 'http://localhost:9999/api/v2/write?bucket=my-bucket&org=my-org&precision=ns',
+                       times: 4, body: request)
+
+      sleep(3)
+    end
+
+    assert_equal('429', error.code)
+
+    assert_requested(:post, 'http://localhost:9999/api/v2/write?bucket=my-bucket&org=my-org&precision=ns',
+                     times: 4, body: request)
+  end
+
+  def test_connection_error
+    error_message = 'Failed to open TCP connection to localhost:9999' \
+        '(Connection refused - connect(2) for "localhost" port 9999)'
+
+    stub_request(:post, 'http://localhost:9999/api/v2/write?bucket=my-bucket&org=my-org&precision=ns')
+      .to_raise(Errno::ECONNREFUSED.new(error_message))
+      .to_raise(Errno::ECONNREFUSED.new(error_message))
+      .to_raise(Errno::ECONNREFUSED.new(error_message))
+      .to_raise(Errno::ECONNREFUSED.new(error_message))
+      .to_raise(Errno::ECONNREFUSED.new(error_message))
+
+    point = InfluxDB2::Point.new(name: 'h2o')
+                            .add_tag('location', 'europe')
+                            .add_field('level', 2.0)
+
+    request = 'h2o,location=europe level=2.0'
+
+    write_options = InfluxDB2::WriteOptions.new(write_type: InfluxDB2::WriteType::BATCHING,
+                                                batch_size: 1, retry_interval: 2_000, max_retries: 3,
+                                                max_retry_delay: 5_000, exponential_base: 2)
+
+    error = assert_raises InfluxDB2::InfluxError do
+      @client.create_write_api(write_options: write_options).write(data: point)
+
+      sleep(0.5)
+
+      assert_requested(:post, 'http://localhost:9999/api/v2/write?bucket=my-bucket&org=my-org&precision=ns',
+                       times: 1, body: request)
+
+      sleep(2)
+
+      assert_requested(:post, 'http://localhost:9999/api/v2/write?bucket=my-bucket&org=my-org&precision=ns',
+                       times: 2, body: request)
+
+      sleep(4)
+
+      assert_requested(:post, 'http://localhost:9999/api/v2/write?bucket=my-bucket&org=my-org&precision=ns',
+                       times: 3, body: request)
+
+      sleep(5)
+
+      assert_requested(:post, 'http://localhost:9999/api/v2/write?bucket=my-bucket&org=my-org&precision=ns',
+                       times: 4, body: request)
+
+      sleep(5)
+
+      assert_requested(:post, 'http://localhost:9999/api/v2/write?bucket=my-bucket&org=my-org&precision=ns',
+                       times: 4, body: request)
+
+      sleep(5)
+    end
+
+    assert_equal('Connection refused - ' + error_message, error.message)
+  end
+
+  def test_write_connection_error
+    stub_request(:post, 'http://localhost:9999/api/v2/write?bucket=my-bucket&org=my-org&precision=ns')
+      .to_raise(Errno::ECONNREFUSED.new(''))
+      .to_raise(Errno::ECONNREFUSED.new(''))
+      .to_return(status: 204)
+
+    point = InfluxDB2::Point.new(name: 'h2o')
+                            .add_tag('location', 'europe')
+                            .add_field('level', 2.0)
+
+    request = 'h2o,location=europe level=2.0'
+
+    write_options = InfluxDB2::WriteOptions.new(write_type: InfluxDB2::WriteType::BATCHING,
+                                                batch_size: 1, retry_interval: 2_000, max_retries: 3,
+                                                max_retry_delay: 5_000, exponential_base: 2)
+
+    @client.create_write_api(write_options: write_options).write(data: point)
+
+    sleep(0.5)
+
+    assert_requested(:post, 'http://localhost:9999/api/v2/write?bucket=my-bucket&org=my-org&precision=ns',
+                     times: 1, body: request)
 
     sleep(2)
 
     assert_requested(:post, 'http://localhost:9999/api/v2/write?bucket=my-bucket&org=my-org&precision=ns',
-                     times: 1, body: request)
+                     times: 2, body: request)
+
+    sleep(4)
+
+    assert_requested(:post, 'http://localhost:9999/api/v2/write?bucket=my-bucket&org=my-org&precision=ns',
+                     times: 3, body: request)
+
+    sleep(5)
+
+    assert_requested(:post, 'http://localhost:9999/api/v2/write?bucket=my-bucket&org=my-org&precision=ns',
+                     times: 3, body: request)
   end
 end